requirements.txt
django-debug-toolbar
mysqlclient==1.3.12
Django==2.0.3
django-json-rpc==0.7.2